+++ examples/relativeParallaxErrorsVsDistance.py
@@ -10,7 +10,11 @@ from pygaia.errors.astrometric import parallaxErrorSkyAvg
 
 from os import environ as env
 from matplotlib import rc
-import argparse
+try:
+  import argparse
+except ImportError:
+  raise ImportError("""The argparse module does not seem to be available.\n Either upgrade to python 2.7
+  or adapt the script to using the optparse module instead.""")
 
 # Configure matplotlib
 rc('text', usetex=True)